Add files via upload
Dijkstra's Algorithm is a classic algorithm used to find the shortest path between a starting node and all other nodes in a weighted graph. It is commonly used in routing and navigation systems, such as GPS, and works on both directed and undirected graphs, provided that the weights (or costs) of edges are non-negative.